requirements.txt
asgiref==3.5.2
asttokens==2.0.8
autopep8==1.7.0
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
executing==1.0.0
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
toml==0.10.2
traitlets==5.3.0
wcwidth==0.2.5
